wip: split and expand tests
